Refine the type of Phi node statements when they are replaced
When we replace a Phi node in `acde_pass!()` due to being able to prove that
one path is always taken, we should request that downstream optimization
passes refine the type of this Phi node in the event that the type of
the replacing value is more narrow than the type of the Phi node itself.

Add a test asserting that the `IR_FLAG_REFINED` flag is added
appropriately.